move pickers to dialog
