+++ src/components/boatdetail.js
@@ -1,24 +1,14 @@
 /* eslint-disable react/jsx-no-target-blank */
 import React, { useState } from 'react';
-import Grid from '@material-ui/core/Grid';
-import Typography from '@material-ui/core/Typography';
 import Paper from '@material-ui/core/Paper';
-import Button from '@material-ui/core/Button';
-import Snackbar from '@material-ui/core/Snackbar';
 import { useTheme } from '@material-ui/core/styles';
 import SwipeableViews from 'react-swipeable-views';
 // import { useInView } from 'react-intersection-observer'
-import { CopyToClipboard } from 'react-copy-to-clipboard';
-import clsx from 'clsx';
 import TabPanel from './tabpanel';
 import ConditionalText from './conditionaltext';
 import SailTable from './sailtable';
-import SmugMugGallery from './smugmuggallery';
 import { feet, price } from './format';
-import ReactFBLike from 'react-fb-like';
-import References from './references';
 import DetailBar from './detailbar';
-import AssignmentIcon from '@material-ui/icons/Assignment';
 
 function m2f(val) {
     if(val) {
@@ -33,18 +23,11 @@ function hullForm(boat) {
   return boat.hull_form.replace(/_/g, ' ');
 }
 
-export default function Boat({ classes, boat, link }) {
+export default function Boat({ classes, boat }) {
   const theme = useTheme();
   const [value, setValue] = useState(0);
-  const [snackBarOpen, setSnackBarOpen] = useState(false);
-  const fixedHeightPaper = clsx(classes.paper, classes.fixedHeight);
-  const fillHeightPaper = clsx(classes.paper, classes.fillHeight);
 
   // TODO const { ref } = useInView({ threshold: 0 });
-  
-  function handleSnackBarClose() {
-    setSnackBarOpen(false);
-  }
   
   const panes = [
     { title: 'Registration and location', children: (
@@ -146,65 +129,19 @@ const engine = {
   };
 
   return (
-    <Grid container spacing={3}>
-      <Grid item xs={12} md={8} lg={9}>
-          <Typography variant="h3" component="h3">{boat.name}</Typography>
-      </Grid>
-      <Grid item xs={12} md={4} lg={3}>
-          <Typography variant="h3" component="h3">{boat.year}</Typography>
-      </Grid>
-      <Grid item xs={12} md={8} lg={9}>
-        <Paper className={fixedHeightPaper}>
-          <SmugMugGallery classes={classes} albumKey={boat.image_key} />
-        </Paper>
-      </Grid>
-      <Grid item xs={12} md={4} lg={3}>
-        <Paper className={fillHeightPaper}>
-          <Typography variant="h4" component="h4">Details</Typography>
-          <ConditionalText value={boat.oga_no} label="OGA no"/>
-          <ConditionalText value={boat.mainsail_type} label="Mainsail"/>
-          <ConditionalText value={boat.rigTypeByRigType && boat.rigTypeByRigType.name} label="Rig"/>
-          <ConditionalText value={boat.home_port} label="Home port or other location"/>
-          <ConditionalText 
-            value={(boat.website)?(<a href={boat.website} rel='noopenner noreferrer' target='_blank'>click here</a>):undefined}
-            label="Website"
-          />
-          <div dangerouslySetInnerHTML={{ __html: boat.short_description }}></div>
-          <References boat={boat}/>
-          <div>
-            <CopyToClipboard text={link} onCopy={() => setSnackBarOpen(true)}>
-              <Button endIcon={<AssignmentIcon/>} size='small' variant='contained' className={classes.button} >
-              Copy page url
-              </Button>
-            </CopyToClipboard>
-              <Snackbar
-              anchorOrigin={{ vertical: 'bottom', horizontal: 'right' }}
-              open={snackBarOpen}
-              autoHideDuration={2000}
-              onClose={handleSnackBarClose}
-              message="URL copied to clipboard."
-              severity="success"
-            />
-          </div>
-          <div>
-            <ReactFBLike href={link} language="en_GB" appId="644249802921642" version="v2.12" />
-          </div>
-          </Paper>
-      </Grid>
-      <Grid item xs={12}>
-          <DetailBar onChange={handleChange} value={value} panes={panes} />
-          <SwipeableViews
-              axis={theme.direction === 'rtl' ? 'x-reverse' : 'x'}
-              index={value}
-              onChangeIndex={handleChangeIndex}
-          >
-          {panes.map((pane, i) => (
-              <TabPanel key={i} value={value} index={i}>
-                  {pane.children}
-              </TabPanel>
-          ))}
-          </SwipeableViews>
-      </Grid>
-    </Grid>
+    <>
+      <DetailBar onChange={handleChange} value={value} panes={panes} />
+      <SwipeableViews
+          axis={theme.direction === 'rtl' ? 'x-reverse' : 'x'}
+          index={value}
+          onChangeIndex={handleChangeIndex}
+      >
+      {panes.map((pane, i) => (
+          <TabPanel key={i} value={value} index={i}>
+              {pane.children}
+          </TabPanel>
+      ))}
+      </SwipeableViews>
+    </>
   );
 }
